// src/api/generated/models/documentRelationships.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { DocumentRelationshipsNodesItem } from './documentRelationshipsNodesItem';
import type { DocumentRelationshipsEdgesItem } from './documentRelationshipsEdgesItem';

export interface DocumentRelationships {
  nodes: DocumentRelationshipsNodesItem[];
  edges: DocumentRelationshipsEdgesItem[];
}